Allow host app & extensions to make use & extend admin's tailwind classes
With this change, the host app and extensions can make use of the
tailwind classes defined in the admin. They can also extend the
tailwind configuration by adding their own classes.

As a requisite for this change, the compiled CSS is now generated
in the host app's assets directory, and tailwindcss-rails needs to
be an explicit dependency for solidus_admin.

The way it works is that we convert the tailwind.config.js file into a
`.js.erb` template. We populate the `content` property by reading from a
`SolidusAdmin::Config#tailwind_content` setting. Through custom
`solidus_admin:tailwindcss:build` and `solidus_admin:tailwindcss:watch` tasks,
we generate the actual tailwind configuration file on the fly.

By default, the `tailwind_content` setting contains both the admin's and the
host app's paths scoped to `solidus_admin` directories, so a host application
is not required to do further configuration. On the other side, extensions can
manually push new paths on an initializer. As solidus_admin's
configuration file requires access to `Rails.root` to create those
defaults, the file must be required after the Rails application is
created.

The solidus_admin's tailwind tasks have been moved to the `lib/tasks` directory
so they can be picked by the `bin/rails` command from within the host app.

When solidus_admin is installed, tailwind is automatically built and the
compiled file is added to the `.gitignore` file. The compilation is also
added as an enhancement for the `assets:precompile` task.