Add column classes to ui/table and use it for the product image column
The product image column requires a specific width. This is now
possible by adding a TailwindCSS class to the column settings
via the `colgroup` and `col` tags.

The class names can be provided to the `colummn` method:

    t.column :image, class_name: "w-1/6"

The keyword argument was called `class_name` and not `class` because
`class` is a reserved keyword in Ruby.

Co-Authored-By: Rainer Dema <rainer.dema@gmail.com>