Set up admin's main navigation routes
We configure the main navigation routes so that the main menu is already
clickable.

For that, we add a `:route` option to the `SolidusAdmin::MainNavItem`.
This can be either a symbol or a proc, and it's resolved by the
components layer when it provides the url helpers as an argument for the
`#path` method. When `:route` is a symbol, it's used as the name of the
helper to be called, while when a proc, it's called with the url helpers
as an argument.

As not all the routes are still available on the new admin, we create a
`SolidusAdmin::UrlHelpersWithFallbacks` module, which acts as a proxy
for the url helpers, and falls back to the old admin ones when
the new ones are not available. When all the work will be done, we can
remove this module and dispatch the new url helpers directly.

We need to accommodate tests for the new `:route` option, and we also
need to mock actual items for the sidebar component to avoid errors because
of the missing routes.

We leave out of this commit mounting the old admin in the dummy
application, which we'll need for feature specs. We might also want to
consider using the url helpers with fallbacks everywhere else.