Implement pagination in Admin Products listing
This commit introduces pagination for the Products listing in the
SolidusAdmin interface with `Geared Pagination`.

The number of products per page is configured through
`SolidusAdmin::Config[:admin_products_per_page]`.

This attribute determines the number of products to display per page
in the admin interface, therefore defining the pagination limit for
the product listing.
By default, the value for this attribute is fetched from the Spree
core configuration and is currently set to display 10 products per page.