Add filter toolbar to table component
This alpha-stage version introduces a filter toolbar to the table component.
Please note that this is an initial representation of the filter architecture,
and it doesn't represent the final version.

Specific changes include:
- Addition of a 'filters' method in the products index component, which takes a
  list of filters.
- Addition of a 'filters' prop in the table component, which takes the list of
  filters and renders them in the filter toolbar.

The goal of this initial implementation is to provide a similar level of features
in the new product index table as the old one, laying the groundwork for
future improvements and updates.

Co-Authored-By: Elia Schito <elia@schito.me>