Add text_field component to solidus_admin
This component is used to render a text field in a form. It leverages
the `type` attribute to also render different input fields, although we
might want to specialize in the future.

It comes in three sizes: small, medium and large. It also supports
rendering a label and a hint, as well as field error messages.

It needs to be rendered in the context of a block yielded by one of
Rails' form helpers, such as `form_for` or `form_with`. This is to
leverage the automatic inferrence of the `name`, `id` and `for`
attributes, therefore avoiding boilerplate. When the given form builder
is bound to a model instance, the error messages will be automatically
extracted. Otherwise, an explicit errors hash needs to be passed.

Ref. #5329