Fix import paths in `solidus_admin/controllers/index.js`
This commit resolves an issue related to the use of relative import paths
in `solidus_admin/controllers/index.js`.

The problem manifested only in the production environment.

Observations:
- In development mode, with the 'compile' setting set to 'true', everything
  was functioning as expected.
- In the production environment, with 'compile' set to 'false' and assets
  being precompiled, the relative paths caused issues.

The root cause appears to be tied to how the production environment handles
module resolution, especially when assets are precompiled.
Changing to absolute paths solves this problem.

After merging this into the `nebulab/admin` branch, the changes will also be
backported into the `admin/alpha/v1` branch.